refactor(skills): consolidate primitive + skill-builder + skill-creator into craft-primitive
Three overlapping skill-management skills (786 lines total) absorbed into one
80-line routing umbrella with 8 focused references, 3 merged scripts, and 1
template. Validator merges best of both (yaml.safe_load + fallback parser,
ALLOWED_PROPERTIES check, JSON output, 500-line hard limit, name-must-match-dir).

Token budget: 786 -> 80 SKILL.md lines + on-demand references.
Budget entries: 3 -> 1 (O(1) instead of O(3)).
