+++ skills/craft-primitive/references/description-patterns.md
