+++ skills/craft-primitive/templates/SKILL-TEMPLATE.md
